### 2023-10-25 #146 @danielfromearth
ncompare
Submitting Author: Daniel Kaufman (@danielfromearth)
Package Name: `ncompare`
One-Line Description of Package: `ncompare` compares two netCDF files at the command line, by generating a report of the matching and non-matching groups, variables, and attributes.
Repository Link:  https://github.com/nasa/ncompare
Version submitted:   1.4.0
Archive: [10.5281/zenodo.10625407](https://zenodo.org/doi/10.5281/zenodo.10625407)
JOSS DOI: [10.21105/joss.06490](https://doi.org/10.21105/joss.06490)
Version accepted: 1.7.2

---

This tool ("ncompare") compares the structure of two Network Common Data Form (NetCDF) files at the command line. It facilitates rapid, human-readable, multi-file evaluation by generating a formatted display of the matching and non-matching groups, variables, and associated metadata between two NetCDF datasets. The user has the option to colorize the terminal output for ease of viewing. As an option, `ncompare` can save comparison reports in text and/or comma-separated value (CSV) formats.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is anyone who manages the generation, manipulation, or validation of netCDF files. This package can be applied to to these netCDF file tasks in any scientific discipline; although it would be most relevant to applications with large multidimensional datasets, e.g., for comparing climate models, for Earth science data reanalyses, and for remote sensing data.

The `ncdiff` function in the `nco` (netCDF Operators) library, as well as `ncmpidiff` and `nccmp`, compute value
differences, but --- as far as we are aware --- do not have a dedicated function to show structural differences between netCDF4 datasets.  Our package, `ncompare` provides a light-weight Python-based tool for rapid **visual** comparisons of group & variable _structures_, _attributes_, and _chunking_.

Pre-submission inquiry #142

- [(NOT YET)] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`.
- [(NOT YET)] The package is deposited in a long-term repository with the DOI:
