### 2023-10-25 #146 @NickleDave
Thanks @danielfromearth for making the full submission!

Here are the initial editor-in-chief checks.

Almost everything we need to start the review is there, but there's one I don't see.

I think you can add this at the top of the single page docs you have now? Basically repeating what you have in the README.

Please let me know if you can add that.
We're ready to find an editor and reviewers once you do.

---
---

- The package imports but there is nothing in the `ncompare` namespace because nothing is imported in ncompare/__init__.py. I understand the package to be designed principally for use through the command-line, so this doesn't affect use
