### 2023-10-26 #146 @danielfromearth
👋 Hi @NickleDave, thanks for the info!
I've now updated the GitHub Pages to include the installation and basic usage section from the `README.md`.
