### 2023-11-22 #146 @tomalrussell
Hi @danielfromearth, thanks for the package submission - we've now identified reviewers and can move on to the next steps..

👋🏻 welcome, @cmarmo and @cmtso! Thank you for volunteering to review
for pyOpenSci ☺️ The review process is outlined in the guides below,
and I'll aim to check in occasionally over the next weeks as we go.

Please get in touch with any questions or concerns! Your review is due in ~3 weeks:  13 December 2023

---

Reviewers: @cmarmo @cmtso
Due date: 13 December 2023
