### 2023-11-24 2dae4cd
add --only-diffs optional arg to cli
