### 2023-11-24 38b344d
check differences before printing a variable to accommodate only-diffs option
