### 2023-11-24 7b50cf2
make only-diffs False by default
