### 2023-11-24 fbe3954
ensure group names, dashes for formatting are preserved even if only-diffs option
