### 2023-11-24 06c0d3a
change type annotation for backwards compatibility
