### 2023-11-24 99aa501
Update out-dated example snippet in README.md
