+++ pyproject.toml
@@ -23,7 +23,7 @@ openpyxl = "^3.1.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.2"
-ruff = ">=0.0.291,<0.1.7"
+ruff = ">=0.0.291,<0.1.8"
 black = "^23.9.1"
 mypy = "^1.5.1"
 pytest-cov = "^4.1.0"
