### 2023-12-08 #146 @cmtso
- [*] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [partial ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README.
- [* ] **Installation instructions:** for the development version of the package and any non-standard dependencies in README.
- [* ] **Vignette(s)** demonstrating major functionality that runs successfully locally.
- [* ] **Function Documentation:** for all user-facing functions.
- [* ] **Examples** for all user-facing functions.

- [* ] Package has a README.md file in the root directory.

- [* ] The package name
- [ *] Badges for:
    - [* ] Current package version (on PyPI / Conda).

- [* ] Short description of package goals.
- [* ] Package installation instructions
- [ *] Any additional setup required to use the package (authentication tokens, etc.)
- [* ] Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.
    - [* ] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)
- [ *] Link to your documentation website.
- [ *] If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.

- [ *] Package documentation is clear and easy to find and use.
- [* ] The need for the package is clear
- [ *] All functions have documentation and associated examples for use
- [ *] The package is easy to install

- [* ] **Installation:** Installation succeeds as documented.
- [* ] **Functionality:** Any functional claims of the software been confirmed.
- [* ] **Performance:** Any performance claims of the software been confirmed.
- [* ] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide).
    - [* ] Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python).

3
---

The package looks very good: package structure, documentation, and tests looks very good. It is a simple-to-use command-line tool based on python to compare the structure of two netCDF file. Note this tool only examines the structure of the files--so if the two files have different values but have the same fields/structure, this tool will not tell them apart.

I have tested the package with all the example tests provided, as well as some of my own data. However, there is no CI in place at the moment and may be something to work on in the future.
