### 2023-12-10 #146 @cmarmo
Hello everybody!

Thanks for giving me the opportunity to review `ncompare`!
Please find my comments following.

- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally: _note that only `ncompare` is exemplified, as the user is not meant to use low-level functions. This comment is applicable to any other documentation reference._

- [ ] **Performance:** Any performance claims of the software been confirmed _(Does not apply)_.
  - [ ] Tests cover essential functions of the package and a reasonable range of inputs and conditions:  _`ncompare/core.py` and `ncompare/printing.py` may benefit of an improved test coverage, both files have a coverage less than 75% and they are the core of the application._
- [x] **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review): _there is no integrated build for the documentation though._
    - [x] Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python): _this is claimed in `pyproject.toml`, but the CI is run against 3.9 only._

2h
---

Below a summary of my comments.

- As `ncompare` is a shell command line it would be informative to have a `--version` option printing the current version.
- I am not used to poetry, the tests can indeed run with `poetry run pytest tests` but `pytest tests` from a python `venv` works too, it could be nice to have this clarified in the `README` as both installations are previously described.
- As python 3.12 is out it could be interesting to have CI for the last python version.
- The license is partially available in the README and as pdf in the `license` directory: as this is a specific NASA license and probably less known by the users, I suggest to have its name specified in the README perhaps with its link at the [OSI website](https://opensource.org/license/nasa1-3-php/), for better discoverability.
- I was able to run `ncompare` on the test data shipped with the package and other files, but the notebook in `ncompare/example` (which is also used in the documentation) uses two files which I was unable to download (permission denied): I strongly recommend to build example with public data, so they can be reproducible.
- I cannot find a documentation build in the CI, am I wrong? This will make easier having consistent notebooks and example for the users.
- The documentation only documents `ncompare`, which is fine because this is the main utility: the CONTRIBUTING file suggests "functions should contain a docstring, though short or trivial function may contain a 1-line docstring", it is unclear to me (as a possible future contributor) how to make the difference between trivial and non-trivial functions, perhaps private and public can be a more consistent definition?
