### 2023-12-12 nasa/ncompare#86 @danielfromearth
Generate coverage report during CI tests
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))

And add a badge — showing the coverage — to the readme
