### 2023-12-12 nasa/ncompare#88 @danielfromearth
Improve test coverage, especially for `ncompare/core.py` and `ncompare/printing.py`
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
