### 2023-12-12 nasa/ncompare#89 @danielfromearth
Add `--version` option to CLI to print current version
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
