### 2023-12-12 nasa/ncompare#90 @danielfromearth
improve readme in a few ways
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))

- [x] clarify poetry vs non-poetry in readme
  - "The tests can indeed run with `poetry run pytest tests` but `pytest tests` from a python venv works too, it could be nice to have this clarified in the README as both installations are previously described."
- [x] license
  - "The license is partially available in the README and as pdf in the license directory: as this is a specific NASA license and probably less known by the users, I suggest to have its name specified in the README perhaps with its link at the [OSI website](https://opensource.org/license/nasa1-3-php/), for better discoverability."
- Add readme badges for
  - [x] Repostatus
  - [x] Supported python versions
  - _References_:
    - https://www.repostatus.org/
    - https://github.com/ResidentMario/missingno/issues/49#issuecomment-362469117
