### 2023-12-12 nasa/ncompare#91 @danielfromearth
try with python 3.12
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
