### 2023-12-12 nasa/ncompare#92 @danielfromearth
Ensure examples utilize publicly accessible data
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
