### 2023-12-12 nasa/ncompare#93 @danielfromearth
Add a documentation build to the CI
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
