### 2023-12-16 cd399d1
add non-poetry pytest test command example to README.md
