### 2023-12-16 58b8893
add unit test for get_vars function
