### 2023-12-16 dc6179c
add test to check error is raised for a non-existent group
