### 2023-12-20 73fd2fc
add tests for printing values to text files and comparison group/var
