### 2023-12-20 5433d53
make sample value printing flexible to different dimensions
