### 2023-12-20 nasa/ncompare#86 @danielfromearth
Perhaps use this Action: https://github.com/marketplace/actions/pytest-coverage-comment
