### 2023-12-27 c74dc4e
get version and use in --version cli arg
