+++ ncompare/console.py
@@ -1,10 +1,13 @@
 #!/usr/bin/env python
 """Command line interface for `ncompare` -- to compare the structure of two NetCDF files."""
 import argparse
+import importlib.metadata
 import sys
 import traceback
 
 from ncompare.core import compare
+
+__version__ = importlib.metadata.version('ncompare')
 
 
 def _cli() -> argparse.Namespace:
@@ -53,7 +56,11 @@ def _cli() -> argparse.Namespace:
     )
 
     parser.add_argument(
-        "--version", action="store_true", default=False, help="Show the current version."
+        "--version",
+        action='version',
+        version=f'%(prog)s {__version__}',
+        default=False,
+        help="Show the current version.",
     )
 
     return parser.parse_args()
