### 2023-12-27 470d855
add cli functional tests for `--version` and `--help`
