### 2023-12-27 1c933e5
consolidate development instructions into new section
