### 2023-12-27 e896e72
improve usage section and remove redundant lines
