### 2023-12-27 927a164
reference NASA OSA license in the contributing section
