+++ README.md
@@ -55,6 +55,11 @@ ncompare S001G01.nc S001G01_SUBSET.nc --file-text subset_comparison.txt
 ## Contributing
 
 Contributions are welcome! For more information see [CONTRIBUTING.md](CONTRIBUTING.md).
+_ncompare_ is licensed under the NASA Open Source Agreement,
+which is included
+[in this repository's license directory](license/LAR-20274-1_ncompare  NetCDF structural comparison tool_NOSA 1.3.pdf)
+and [on the Open Source Initiative website](https://opensource.org/license/nasa1-3-php/).
+
 
 ### Developing
 
