+++ README.md
@@ -2,7 +2,7 @@
 _____
 
 <a href="https://pypi.org/project/ncompare" target="_blank">
-    <img src="https://img.shields.io/pypi/v/ncompare?color=%2334D058&label=pypi%20package" alt="Package version">
+    <img src="https://img.shields.io/pypi/v/ncompare?color=%2334D058label=pypi%20package" alt="Package version">
 </a>
 <a href="https://pypi.org/project/ncompare/" target="_blank">
     <img src="https://img.shields.io/pypi/pyversions/ncompare.svg" alt="Python Versions">
@@ -15,6 +15,9 @@ _____
 </a>
 <a href="http://mypy-lang.org/" target="_blank">
     <img src="http://www.mypy-lang.org/static/mypy_badge.svg" alt="Mypy checked">
+</a>
+<a href="https://www.repostatus.org/#active" target="_blank">
+    <img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active – The project has reached a stable, usable state and is being actively developed">
 </a>
 
 Compare the structure of two NetCDF files at the command line.
