### 2023-12-27 03b2a03
remove version from argparser namespace to prevent exception
