### 2023-12-27 0d9fb89
add golden files and tests for csv and xlsx outputs
