### 2023-12-27 31ceda4
modify cli parser to enable unit testing
