### 2023-12-27 1370700
add github actions test coverage badge step and .images directory
