+++ .github/workflows/reusable_run_tests.yml
@@ -41,6 +41,33 @@ jobs:
         run: |
           poetry run pytest --cov=ncompare --cov-report=xml:build/reports/coverage${{ matrix.python-version }}.xml --cov-report=html:build/reports/coverage${{ matrix.python-version }}.html tests/
 
+      - uses: tj-actions/coverage-badge-py@v2
+        id: coverage-badge-py
+        with:
+          # Output path (str) to write the coverage badge. Default: "coverage.svg".
+          output: '.images/coverage.svg'
+
+      - name: Verify Changed files
+        uses: tj-actions/verify-changed-files@v16
+        id: verify-changed-files
+        with:
+          files: .images/coverage.svg
+
+      - name: Commit files
+        if: steps.verify-changed-files.outputs.files_changed == 'true'
+        run: |
+          git config --local user.email "ncompare@noreply.github.com"
+          git config --local user.name "ncompare bot"
+          git add .images/coverage.svg
+          git commit -m "Updated coverage.svg"
+
+      - name: Push changes
+        if: steps.verify-changed-files.outputs.files_changed == 'true'
+        uses: ad-m/github-push-action@master
+        with:
+          github_token: ${{ secrets.github_token }}
+          branch: ${{ github.ref }}
+
       - name: Archive code coverage report (xml)
         if: ${{ always() }}
         uses: actions/upload-artifact@v4
