+++ .github/workflows/reusable_run_tests.yml
@@ -39,7 +39,29 @@ jobs:
 
       - name: Run tests with coverage
         run: |
-          poetry run pytest --cov=ncompare --cov-report=xml:build/reports/coverage${{ matrix.python-version }}.xml --cov-report=html:build/reports/coverage${{ matrix.python-version }}.html tests/
+          poetry run pytest --cov=ncompare --cov-report=xml:build/reports/coverage${{ matrix.python-version }}.xml --cov-report=html:build/reports/coverage${{ matrix.python-version }}.html --cov-report= tests/
+
+      - name: Archive code coverage report (xml)
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: code coverage report (xml)
+          path: build/reports/coverage${{ matrix.python-version }}.xml
+
+      - name: Archive code coverage report (HTML)
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: code coverage report (HTML)
+          path: build/reports/coverage${{ matrix.python-version }}.html
+
+
+  coverage:
+    runs-on: ubuntu-latest
+    needs: build_and_test
+
+    steps:
+      - uses: actions/checkout@v4
 
       - uses: tj-actions/coverage-badge-py@v2
         id: coverage-badge-py
@@ -67,17 +89,3 @@ jobs:
         with:
           github_token: ${{ secrets.github_token }}
           branch: ${{ github.ref }}
-
-      - name: Archive code coverage report (xml)
-        if: ${{ always() }}
-        uses: actions/upload-artifact@v4
-        with:
-          name: code coverage report (xml)
-          path: build/reports/coverage.xml
-
-      - name: Archive code coverage report (HTML)
-        if: ${{ always() }}
-        uses: actions/upload-artifact@v4
-        with:
-          name: code coverage report (HTML)
-          path: build/reports/coverage.html
