+++ .github/workflows/reusable_run_tests.yml
@@ -57,6 +57,7 @@ jobs:
 
 
   coverage:
+    name: Coverage
     runs-on: ubuntu-latest
     needs: build_and_test
 
