+++ .github/workflows/reusable_run_tests.yml
@@ -41,52 +41,57 @@ jobs:
         run: |
           poetry run pytest --cov=ncompare --cov-report=xml:build/reports/coverage${{ matrix.python-version }}.xml --cov-report=html:build/reports/coverage${{ matrix.python-version }}.html --cov-report= tests/
 
-      - name: Archive code coverage report (xml)
-        if: ${{ always() }}
-        uses: actions/upload-artifact@v4
+      - uses: codecov/codecov-action@v3
         with:
-          name: code coverage report (xml)
-          path: build/reports/coverage${{ matrix.python-version }}.xml
+          files: build/reports/coverage${{ matrix.python-version }}.xml
+          verbose: true # optional (default = false)
 
-      - name: Archive code coverage report (HTML)
-        if: ${{ always() }}
-        uses: actions/upload-artifact@v4
-        with:
-          name: code coverage report (HTML)
-          path: build/reports/coverage${{ matrix.python-version }}.html
-
-
-  coverage:
-    name: Coverage
-    runs-on: ubuntu-latest
-    needs: build_and_test
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: tj-actions/coverage-badge-py@v2
-        id: coverage-badge-py
-        with:
-          # Output path (str) to write the coverage badge. Default: "coverage.svg".
-          output: '.images/coverage.svg'
-
-      - name: Verify Changed files
-        uses: tj-actions/verify-changed-files@v16
-        id: verify-changed-files
-        with:
-          files: .images/coverage.svg
-
-      - name: Commit files
-        if: steps.verify-changed-files.outputs.files_changed == 'true'
-        run: |
-          git config --local user.email "ncompare@noreply.github.com"
-          git config --local user.name "ncompare bot"
-          git add .images/coverage.svg
-          git commit -m "Updated coverage.svg"
-
-      - name: Push changes
-        if: steps.verify-changed-files.outputs.files_changed == 'true'
-        uses: ad-m/github-push-action@master
-        with:
-          github_token: ${{ secrets.github_token }}
-          branch: ${{ github.ref }}
+#      - name: Archive code coverage report (xml)
+#        if: ${{ always() }}
+#        uses: actions/upload-artifact@v4
+#        with:
+#          name: code coverage report (xml)
+#          path: build/reports/coverage${{ matrix.python-version }}.xml
+#
+#      - name: Archive code coverage report (HTML)
+#        if: ${{ always() }}
+#        uses: actions/upload-artifact@v4
+#        with:
+#          name: code coverage report (HTML)
+#          path: build/reports/coverage${{ matrix.python-version }}.html
+#
+#
+#  coverage:
+#    name: Coverage
+#    runs-on: ubuntu-latest
+#    needs: build_and_test
+#
+#    steps:
+#      - uses: actions/checkout@v4
+#
+#      - uses: tj-actions/coverage-badge-py@v2
+#        id: coverage-badge-py
+#        with:
+#          # Output path (str) to write the coverage badge. Default: "coverage.svg".
+#          output: '.images/coverage.svg'
+#
+#      - name: Verify Changed files
+#        uses: tj-actions/verify-changed-files@v16
+#        id: verify-changed-files
+#        with:
+#          files: .images/coverage.svg
+#
+#      - name: Commit files
+#        if: steps.verify-changed-files.outputs.files_changed == 'true'
+#        run: |
+#          git config --local user.email "ncompare@noreply.github.com"
+#          git config --local user.name "ncompare bot"
+#          git add .images/coverage.svg
+#          git commit -m "Updated coverage.svg"
+#
+#      - name: Push changes
[…]
