### 2023-12-28 036bd5f
add python version 3.12 to test runner
