### 2023-12-28 nasa/ncompare#92 @danielfromearth
This public data can be included in the automatically built example notebook.
