### 2024-01-03 69100b2
add quarto yml config
