### 2024-01-03 73ddd94
move quarto config to docs subdirectory
