+++ .github/workflows/quarto-publish.yml
@@ -27,5 +27,6 @@ jobs:
       uses: quarto-dev/quarto-actions/publish@v2
       with:
         target: gh-pages
+        path: docs
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # this secret is always available for github actions
