### 2024-01-03 a687991
move quarto config to docs subdirectory
