### 2024-01-03 e087071
rework docs for rtd
