### 2024-01-03 d72efdd
rework docs for rtd
