+++ .github/workflows/quarto-publish.yml
@@ -1,32 +1,0 @@
-name: Render and deploy quarto files
-on:
-  push:
-    branches: [develop]
-  pull_request:
-    branches: [develop]
-
-jobs:
-  quarto-render-and-deploy:
-    runs-on: ubuntu-latest
-    #    defaults:
-    #      run:
-    #        shell: bash -l {0}
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Set up Quarto
-      uses: quarto-dev/quarto-actions/setup@v2
-      with:
-        # To install LaTeX to build PDF book
-        tinytex: true
-        # uncomment below and fill to pin a version
-        # version: 0.9.600
-
-    - name: Publish to GitHub Pages (and render)
-      uses: quarto-dev/quarto-actions/publish@v2
-      with:
-        target: gh-pages
-        path: docs
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # this secret is always available for github actions
