+++ mkdocs.yml
@@ -100,4 +100,3 @@ markdown_extensions:
 watch:
   - docs
   - ncompare
-  - notebooks
