### 2024-01-04 2a2bb67
add mkdocs dependencies to poetry dev group and update
