### 2024-01-04 8dd19f8
fix missing nav links
