### 2024-01-04 4f0ac6b
turn mkdocs-jupyter to False
