+++ mkdocs.yml
@@ -40,7 +40,7 @@ plugins:
             show_root_heading: false
             show_source: false
   - mkdocs-jupyter:
-      execute: True
+      execute: False
       ignore:
         #        - "tutorials/dask.ipynb"
         - "**/.ipynb_checkpoints/*"
