### 2024-01-05 aff4302
convert doc markdowns to symlinks
