+++ docs/CODE_OF_CONDUCT.md
@@ -1,1 +1,1 @@
-../CODE_OF_CONDUCT.md
+CODE_OF_CONDUCT.md
