### 2024-01-10 nasa/ncompare#86 @danielfromearth
Use Codecov, like other NASA projects (https://codecov.io/gh/nasa).
