### 2024-01-10 aef5cdd
add codecov step to tests workflow
