+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [pull-request/99](https://github.com/nasa/ncompare/pull/99): Improve readme in a few ways (e.g., license, badges)
 - [pull-request/106](https://github.com/nasa/ncompare/pull/106): Use ReadTheDocs instead of GitHub Pages for documentation
+- [pull-request/113](https://github.com/nasa/ncompare/pull/113): Add codecov step to tests workflow
 ### Deprecated
 ### Removed
 - fixed bug related to extra argument from command line
