+++ .github/workflows/pr-rtd-link.yml
@@ -18,4 +18,4 @@ jobs:
     steps:
       - uses: "readthedocs/actions/preview@v1"
         with:
-          project-slug: "earthaccess"
+          project-slug: "ncompare"
