+++ .github/workflows/build-pipeline.yml
@@ -1,4 +1,4 @@
-name: Lint and Test
+name: CI
 
 # Controls when the workflow will run
 on:
