### 2024-01-11 #146 @danielfromearth
Hi folks, a quick update: I'm wrapping up some final tweaks to respond to the review comments, and will also hopefully be able to add a JOSS paper to the repo next week.
