### 2024-01-12 098658b
Bump abatilo/actions-poetry from 2.4.0 to 3.0.0
Bumps [abatilo/actions-poetry](https://github.com/abatilo/actions-poetry) from 2.4.0 to 3.0.0.
- [Release notes](https://github.com/abatilo/actions-poetry/releases)
- [Changelog](https://github.com/abatilo/actions-poetry/blob/master/.releaserc)
- [Commits](https://github.com/abatilo/actions-poetry/compare/v2.4.0...v3.0.0)

---
updated-dependencies:
- dependency-name: abatilo/actions-poetry
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
