+++ .github/workflows/build-pipeline.yml
@@ -25,7 +25,7 @@ jobs:
         python-version: 3.9
         
     - name: Install Poetry
-      uses: abatilo/actions-poetry@v2.4.0
+      uses: abatilo/actions-poetry@v3.0.0
       with:
         poetry-version: 1.3.2
         
