### 2024-01-16 nasa/ncompare#118 @danielfromearth
Add citation file
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
