### 2024-01-16 nasa/ncompare#122 @danielfromearth
Add manuscript for Journal of Open Source Software (JOSS)
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))
