### 2024-01-16 9c20671
tweak contributing wording
