+++ CONTRIBUTING.md
@@ -21,14 +21,14 @@ Thank you for contributing to _ncompare_!
 
 ### Reporting Bugs
 
-Bugs can be reported by creating a new GitHub issue in this repository. 
-Please use the `bug` label on any bug issue created. Provide detailed 
-instructions on how to reproduce the bug if possible. Include the version 
+Bugs can be reported by creating a new GitHub issue in this repository.
+Please use the `bug` label on any bug issue created. Provide detailed
+instructions on how to reproduce the bug if possible. Include the version
 of _ncompare_ that produced this bug and specific error messages.
 
 ### Feature/Enhancement Requests
 
-Feature/enhancement requests are tracked with GitHub issues. Please use 
+Feature/enhancement requests are tracked with GitHub issues. Please use
 an appropriate label:
 
 - `documentation`
@@ -39,13 +39,13 @@ an appropriate label:
     - Improvement of an existing feature
 - `new-feature`
     - An entirely new feature
-    
-Feature and enhancement requests should contain a detailed description 
+
+Feature and enhancement requests should contain a detailed description
 of the desired behavior and rationale.
 
 ### Pull Requests
 
-Please ensure all pull requests follow the PR 
+Please ensure all pull requests follow the PR
 [template](/.github/pull_request_template.md). This is to ensure
 
 * A reviewer understands what is being changed
@@ -53,9 +53,9 @@ Please ensure all pull requests follow the PR
 * A reviewer understands how the changes have been verified
 * A reviewer is confident the changes work as expected and won't break existing functionality.
 
-The name of the pull request should be meaningful, and if the pull 
-request is addressing a GitHub issue, should match the name of the 
-GitHub issue. 
+The name of the pull request should be meaningful, and if the pull
+request is addressing a GitHub issue, should match the name of the
+GitHub issue.
 
 If any performance improvements are being made, please include graphs or charts.
 
@@ -66,7 +66,7 @@ If any performance improvements are being made, please include graphs or charts.
 - `main` (protected)
     - Matches the version currently in prod
 - `release/#.#.#`
-    - Release branch being considered for delivery to prod. 
+    - Release branch being considered for delivery to prod.
     - Matches the version currently in test
 - `feature/issue-#`
     - Work for enhancements and new features should be done in a branch with this naming convention
@@ -76,18 +76,18 @@ If any performance improvements are being made, please include graphs or charts.
     - The issue number should match the associated GitHub issue number
 - `hotfix/issue-#` or `hotfix/short-fix-description`
     - Rare/special case to address a special anomaly.
-    - The issue number should match the associated GitHub issue number, 
-    unless no such issue exists. If not, use a short description of the 
+    - The issue number should match the associated GitHub issue number,
+    unless no such issue exists. If not, use a short description of the
     issue e.g. `hotfix/fix-request-url`
 
 ### Changelog
 
-_Ncompare_ maintains a [changelog](CHANGELOG.md). See 
-[Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for more 
+_Ncompare_ maintains a [changelog](CHANGELOG.md). See
+[Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for more
 information.
 
-The top of the changelog will contain an `[Unreleased]` section. Add 
-any changes made in the current PR to the appropriate section of the 
+The top of the changelog will contain an `[Unreleased]` section. Add
+any changes made in the current PR to the appropriate section of the
 changelog.
 
 Please keep changelog messages relatively short (1-2 sentences).
@@ -96,48 +96,48 @@ Please keep changelog messages relatively short (1-2 sentences).
 
 In most cases, unit tests should be either created or updated in every PR.
 
[…]
