### 2024-01-16 c6142fe
tweak contributing wording
