+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [pull-request/113](https://github.com/nasa/ncompare/pull/113): Add codecov step to tests workflow
 - [issue/88](https://github.com/nasa/ncompare/issues/88): Improve test coverage, especially for ncompare/core.py and ncompare/printing.py
 - [issue/92](https://github.com/nasa/ncompare/issues/92): Ensure examples utilize publicly accessible data
+- [pull-request/121](https://github.com/nasa/ncompare/pull/121): Disable text wrapping to properly show ncompare output in notebook example
 - Tweaked wording regarding docstrings in the contributing guide
 ### Deprecated
 ### Removed
