### 2024-01-16 #146 @danielfromearth
My responses to the comments and blank check-boxes from @cmtso's review can be found next to the green checkmarks (✅) below:

> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  (partial) **A statement of need** clearly stating problems the software is designed to solve and its target audience in README.
> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING.
> * [ ]  **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `pyproject.toml` file or elsewhere.
>

- ✅ A statement of need can now be found in the [`README.md`](https://github.com/nasa/ncompare/blob/main/README.md)
- ✅ Community guidelines are now present in the [`CONTRIBUTING.md`](https://github.com/nasa/ncompare/blob/main/CONTRIBUTING.md) file.
- ✅ [pyproject.toml](https://github.com/nasa/ncompare/blob/main/pyproject.toml) now includes author, author email, and url.

> The README should include, from top to bottom:
> * [x]  Badges for:
>
>   * [ ]  Continuous integration and test coverage,
>   * [ ]  Docs building (if you have a documentation website),
>   * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,
>   * [ ]  Python versions supported,

- ✅ Badges for CI and test coverage, docs, repostatus, and python versions have been added to the [`README.md`](https://github.com/nasa/ncompare/blob/main/README.md)

> * [ ]  Citation information

- ✅ This has been added here: [`CITATION.cff`](https://github.com/nasa/ncompare/blob/main/CITATION.cff)

> #### Functionality
> * [ ]  **Automated tests:**
>
>   * [ ]  All tests pass on the reviewer's local machine for the package version submitted by the author. Ideally this should be a tagged version making it easy for reviewers to install.
>   * [ ]  Tests cover essential functions of the package and a reasonable range of inputs and conditions.
> * [ ]  **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)

- ✅ Automated tests are part of the main Continuous Integration workflow defined [here](https://github.com/nasa/ncompare/blob/aba145dab632b85d456b37f79f193ac80a7255cf/.github/workflows/build-pipeline.yml#L14), and these tests are run as part of GitHub Actions shown [here](https://github.com/nasa/ncompare/actions/workflows/build-pipeline.yml).  The tests cover essential functions, and test coverage can now be [seen via Codecov](https://app.codecov.io/gh/nasa/ncompare).

> * [x]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide).
>   A few notable highlights to look at:
>
>   * [x]  Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python).
>   * [ ]  Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

- ✅ The code format is further standardized and uses `ruff` for linting, as defined [here] (https://github.com/nasa/ncompare/blob/3f0d99f30051bac305ee1fdb26dbb9555fcb72f8/.github/workflows/reusable_run_tests.yml#L36) in the CI workflow.

> #### For packages also submitting to JOSS
> * [ ]  The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).
>
> _Note:_ Be sure to check this carefully, as JOSS's submission requirements and scope differ from pyOpenSci's in terms of what types of packages are accepted.
>
> The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
> * [ ]  **A short summary** describing the high-level functionality of the software
> * [ ]  **Authors:** A list of authors with their affiliations
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

- ✅ The JOSS manuscript with those sections has been added to the `joss-paper` branch in [this `paper.md` file](https://github.com/nasa/ncompare/blob/joss-paper/paper.md).

> ## Estimated hours spent reviewing:
> 3
> #### Review Comments
> The package looks very good: package structure, documentation, and tests looks very good. It is a simple-to-use command-line tool based on python to compare the structure of two netCDF file. Note this tool only examines the structure of the files--so if the two files have different values but have the same fields/structure, this tool will not tell them apart.
>
> I have tested the package with all the example tests provided, as well as some of my own data. However, there is no CI in place at the moment and may be something to work on in the future.

- ✅ Thank you very much @cmtso for your comments and the time and attention you devoted to this review!
