### 2024-01-16 #146 @danielfromearth
My responses to the comments and blank check-boxes from @cmarmo's review can be found next to the green checkmarks (✅) below:

> The README should include, from top to bottom:
> * [ ]  Badges for:
>
>   * [ ]  Continuous integration and test coverage,
>   * [ ]  Docs building (if you have a documentation website),
>   * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,
>   * [ ]  Python versions supported,
>   * [x]  Current package version (on PyPI / Conda).

- ✅ Badges for CI and test coverage, docs, repostatus, and python versions have been added to the [`README.md`](https://github.com/nasa/ncompare/blob/main/README.md)

> * [ ]  Link to your documentation website.

- ✅ A link is now in the `docs: passing` badge.

> * [ ]  Citation information

- ✅ This has been added here: [`CITATION.cff`](https://github.com/nasa/ncompare/blob/main/CITATION.cff)

> #### Functionality
> * [ ]  **Automated tests:**
>   * [ ]  Tests cover essential functions of the package and a reasonable range of inputs and conditions:  _`ncompare/core.py` and `ncompare/printing.py` may benefit of an improved test coverage, both files have a coverage less than 75% and they are the core of the application._

- ✅ Automated tests are part of the main Continuous Integration workflow defined [here](https://github.com/nasa/ncompare/blob/aba145dab632b85d456b37f79f193ac80a7255cf/.github/workflows/build-pipeline.yml#L14), and these tests are run as part of GitHub Actions shown [here](https://github.com/nasa/ncompare/actions/workflows/build-pipeline.yml).  The tests cover essential functions, and the improved test coverage (`core.py` is now at 90% and `printing.py` at 87%) can now be [seen via Codecov](https://app.codecov.io/gh/nasa/ncompare).

> #### For packages also submitting to JOSS
> * [ ]  The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).
>
> _Note:_ Be sure to check this carefully, as JOSS's submission requirements and scope differ from pyOpenSci's in terms of what types of packages are accepted.
>
> The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
> * [ ]  **A short summary** describing the high-level functionality of the software
> * [ ]  **Authors:** A list of authors with their affiliations
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

- ✅ The JOSS manuscript with those sections has been added to the `joss-paper` branch in [this `paper.md` file](https://github.com/nasa/ncompare/blob/joss-paper/paper.md).

> ## Estimated hours spent reviewing:
> 2h
> #### Review Comments
> Below a summary of my comments.
>
> * As `ncompare` is a shell command line it would be informative to have a `--version` option printing the current version.

- ✅ This option has now been added to the command line interface.

> * I am not used to poetry, the tests can indeed run with `poetry run pytest tests` but `pytest tests` from a python `venv` works too, it could be nice to have this clarified in the `README` as both installations are previously described.

- ✅ This is a good suggestion, and it is now part of the README.

> * As python 3.12 is out it could be interesting to have CI for the last python version.

- ✅ Done.

> * The license is partially available in the README and as pdf in the `license` directory: as this is a specific NASA license and probably less known by the users, I suggest to have its name specified in the README perhaps with its link at the [OSI website](https://opensource.org/license/nasa1-3-php/), for better discoverability.

- ✅ The suggested information has been added to the "Contributing" section of the README.md

> * I was able to run `ncompare` on the test data shipped with the package and other files, but the notebook in `ncompare/example` (which is also used in the documentation) uses two files which I was unable to download (permission denied): I strongly recommend to build example with public data, so they can be reproducible.

- ✅ The notebook has been updated to use publicly accessible data files.

> * I cannot find a documentation build in the CI, am I wrong? This will make easier having consistent notebooks and example for the users.

- ✅ Documentation has been changed from GitHub pages to use ReadTheDocs, and this is now part of the CI build.

> * The documentation only documents `ncompare`, which is fine because this is the main utility: the CONTRIBUTING file suggests "functions should contain a docstring, though short or trivial function may contain a 1-line docstring", it is unclear to me (as a possible future contributor) how to make the difference between trivial and non-trivial functions, perhaps private and public can be a more consistent definition?

- ✅  I understand how this may lead to some uncertainty, although I don't think complicated private functions should have 1-line docstrings either. There will unavoidably be some subjectivity. Therefore, this description has been modified to "...private functions (that are exceptionally short)...".

- ✅ Thank you very much @cmarmo for these additional comments and the time and attention you devoted to reviewing `ncompare`!
