### 2024-01-16 c7dc881
update pyproject.toml classifiers
