+++ pyproject.toml
@@ -7,7 +7,16 @@ readme = "README.md"
 repository = "https://github.com/nasa/ncompare"
 license = "license/LAR-20274-1_ncompare  NetCDF structural comparison tool_NOSA 1.3.pdf"
 classifiers = [
-    "Topic :: Software Development :: Libraries :: Python Modules"
+    "Environment :: Console",
+    "Intended Audience :: Science/Research",
+    "Operating System :: Microsoft :: Windows",
+    "Operating System :: MacOS",
+    "Operating System :: POSIX :: Linux",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 
 [tool.poetry.scripts]
