### 2024-01-16 5dbe4df
updated CHANGELOG.md and CITATION.cff
