+++ CITATION.cff
@@ -18,7 +18,7 @@ keywords:
 url: "https://ncompare.readthedocs.io"
 repository-code: "https://github.com/nasa/ncompare"
 
-version: 1.6.1
+version: 1.6.2
 date-released: 2024-01-16
 
 authors:
