### 2024-01-16 #146 @danielfromearth
Hi @tomalrussell, thanks for shepherding this submission thus far.

We have made changes to the repository to address the reviewer comments, and responded directly to the review points above.  In addition, please notice that we have now included a JOSS manuscript in the repository (in the `joss-paper` branch). Cheers.
