### 2024-01-17 #146 @tomalrussell
Thanks for the super clear responses @danielfromearth !

@cmarmo and @cmtso - over to you to respond to the changes. Discussion is encouraged, we're in the ["review follow-up"](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#review-follow-up) phase.

Could you pay particular attention to the `For packages also submitting to JOSS` section, now that the manuscript is up on the repository ([`paper.md`](https://github.com/nasa/ncompare/blob/joss-paper/paper.md) and [`paper.bib`](https://github.com/nasa/ncompare/blob/joss-paper/paper.bib))? It may be helpful to refer to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper) for more detail and background on the expectations here.

As ever, do ask or drop me a line if you need anything ☺️

Responses due in ~2 weeks: 31 January.
