### 2024-01-22 #146 @cmarmo
Hello @danielfromearth , thank you for all your work and your answers to my comments!
I have checked all your improvements and I have still some comments about the documentation:
- in the `README.md` the link to the license inside the package is not rendered (I cannot understand why at first sight... )
- In the `README.md` and the documentation the link to the example jupyter notebook gives a 404 error
- The file https://gpm1.gesdisc.eosdis.nasa.gov/data/GPM_L3/GPM_3IMERGDL.06/2024/01/3B-DAY-L.MS.MRG.3IMERG.20240110-S000000-E235959.V06.nc4 used in the notebook still asks me for user and password.... am I missing something?

I'm going to check the paper for JOSS in the next week.... sorry for the two steps answer....
