### 2024-01-23 eb16d95
Bump xarray from 2023.12.0 to 2024.1.0
Bumps [xarray](https://github.com/pydata/xarray) from 2023.12.0 to 2024.1.0.
- [Release notes](https://github.com/pydata/xarray/releases)
- [Changelog](https://github.com/pydata/xarray/blob/main/HOW_TO_RELEASE.md)
- [Commits](https://github.com/pydata/xarray/compare/v2023.12.0...v2024.01.0)

---
updated-dependencies:
- dependency-name: xarray
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
