### 2024-01-23 nasa/ncompare#134 @danielfromearth
Fix repository links in the README
(As part of `ncompare`'s [PyOpenSci review (issue #146)](https://github.com/pyOpenSci/software-submission/issues/146))

The following two broken links, mentioned in [this review comment](https://github.com/pyOpenSci/software-submission/issues/146#issuecomment-1904841655), should be fixed:
> * in the `README.md` the link to the license inside the package is not rendered (I cannot understand why at first sight... )
> * In the `README.md` and the documentation the link to the example jupyter notebook gives a 404 error
