### 2024-01-26 fe9f92d
fix example notebook link to point to readthedocs link
