### 2024-01-27 #146 @cmarmo
Hello @danielfromearth, I realized that you are merging your modifications in a `develop` branch rather than in `main`.
I see that the link to the notebook and to the example files are fixed: thank you!
It looks like the link to the local version of the license in the readme  still contains spaces and it is still not rendered.

@tomalrussell after this nit fixing I consider the author has responded to my review and made changes to my satisfaction. I recommend approving this package.

I have also checked the paper for JOSS: I'm copying here the checklist related to JOSS.

- [x] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements): I believe `ncompare` can be considered as  a tool to "extract knowledge from large data sets". However, note that JOSS states that "single-function packages are not acceptable." It is not clear to me how to interpret "single-function" and if `ncompare` can be considered as one.

- [ ] **Authors:** A list of authors with their affiliations: the affiliation of the author is missing
