### 2024-01-29 5833908
Bump markdown-callouts from 0.3.0 to 0.4.0
Bumps [markdown-callouts](https://github.com/oprypin/markdown-callouts) from 0.3.0 to 0.4.0.
- [Release notes](https://github.com/oprypin/markdown-callouts/releases)
- [Commits](https://github.com/oprypin/markdown-callouts/compare/v0.3.0...v0.4.0)

---
updated-dependencies:
- dependency-name: markdown-callouts
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
