+++ pyproject.toml
@@ -38,7 +38,7 @@ mypy = "^1.5.1"
 pytest-cov = "^4.1.0"
 mkdocs = "^1.5.3"
 mkdocs-material = "^9.5.3"
-markdown-callouts = "^0.3.0"
+markdown-callouts = ">=0.3,<0.5"
 markdown-include = "^0.8.1"
 mkdocstrings = "^0.24.0"
 mkdocs-jupyter = "^0.24.6"
