### 2024-01-30 #146 @cmtso
- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements):  I think this is a hugely useful tool to the research community. At the moment, it may be regarded as a minor 'utility' package. Perhaps a clearer roadmap of future feature extensions to cover a slightly larger (and more specific) scope and providing research use case in the paper will help with the JOSS submission.

- [ *] **A short summary** describing the high-level functionality of the software
- [ *] **Authors:** A list of authors with their affiliations
- [ *] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- [ *] **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

#### Final approval (post-review)

- [ *] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
