### 2024-02-02 #146 @tomalrussell
Hi @danielfromearth, thanks for the thorough response.

On the question of scope for JOSS - in the end this is up to the JOSS editorial team, but do refer to this pyOpenSci review (as for example in this recent JOSS submission [openjournals/joss-reviews#5332](https://github.com/openjournals/joss-reviews/issues/5332#issuecomment-1494828519)).  I think you have a strong case for `ncompare` as enabling tool that extracts useful data from large datasets. Revising the manuscript as you describe sounds worthwhile. I'm happy to recommend it for submission when you're ready.

On the last few nits - they all look to be resolved on `develop`, and merging and creating a release should be part of the wrap-up, so with that -

---

🎉 `ncompare` has been approved by pyOpenSci! Thank you @danielfromearth for submitting `ncompare` and many thanks again to @cmarmo and @cmtso for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of `ncompare`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/146)`.

### JOSS submission

Here are the next steps:

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

Please complete the final steps to wrap up this review. @tomalrussell reminder to do the following:

---
