+++ CITATION.cff
@@ -18,8 +18,8 @@ keywords:
 url: "https://ncompare.readthedocs.io"
 repository-code: "https://github.com/nasa/ncompare"
 
-version: 1.6.2
-date-released: 2024-01-16
+version: 1.7.1
+date-released: 2024-02-06
 
 authors:
 - family-names: "Baskin"
