+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.2]
+### Added
+### Changed
+- add pyOpenSci badge to readme
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.7.1]
 ### Added
 ### Changed
