### 2024-02-06 #146 @danielfromearth
> ## Author Wrap Up Tasks
> There are a few things left to do to wrap up this submission:
>
> * [x]  Activate [Zenodo](https://zenodo.org/) watching the repo if you haven't already done so.
> * [x]  Tag and create a release to create a Zenodo version and DOI.
> * [x]  Add the badge for pyOpenSci peer-review to the README.md of `ncompare`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/146)`.
> * [x]  Please fill out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7). All maintainers and reviewers should fill this out.
>
> ### JOSS submission
> Here are the next steps:
>
> * [ ]   Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.
> * [ ]  Login to the JOSS website and fill out the JOSS submission form using your Zenodo DOI. **When you fill out the form, be sure to mention and link to the approved pyOpenSci review.** JOSS will tag your package for expedited review if it is already pyOpenSci approved.
> * [ ]  Wait for a JOSS editor to approve the presubmission (which includes a scope check).
> * [ ]  Once the package is approved by JOSS, you will be given instructions by JOSS about updating the citation information in your README file.
> * [ ]  When the JOSS review is complete, add a comment to your review in the pyOpenSci software-review repo here that it has been approved by JOSS. An editor will then add the JOSS-approved label to this issue.

✅ `ncompare` has been archived on Zenodo, with this DOI: [10.5281/zenodo.10625407](https://zenodo.org/doi/10.5281/zenodo.10625407)
