+++ pyproject.toml
@@ -32,7 +32,7 @@ openpyxl = "^3.1.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.4.2,<9.0.0"
-ruff = ">=0.0.291,<0.2.1"
+ruff = ">=0.0.291,<0.2.2"
 black = ">=23.9.1,<25.0.0"
 mypy = "^1.5.1"
 pytest-cov = "^4.1.0"
