+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Compare the structure of two NetCDF files at the command line"
 authors = ["Daniel Kaufman <daniel.kaufman@nasa.gov>"]
 readme = "README.md"
 repository = "https://github.com/nasa/ncompare"
-license = "license/LAR-20274-1_ncompare  NetCDF structural comparison tool_NOSA 1.3.pdf"
+license = "license/LAR-20274-1_ncompare_NetCDF_structural_comparison_tool_NOSA_1.3.pdf"
 classifiers = [
     "Environment :: Console",
     "Intended Audience :: Science/Research",
