+++ pyproject.toml
@@ -41,6 +41,7 @@ markdown-callouts = "^0.4.0"
 markdown-include = "^0.8.1"
 mkdocstrings = "^0.24.1"
 mkdocs-jupyter = "^0.24.6"
+mkdocs-material = "^9.5.12"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
