### 2024-06-11 #146 @lwasser
hey team. i'm checking in on this issue. did this issue / package get fastraacked through joss? if it did we need to do a few followup steps!

1. cross link this issue with the joss review issue. i see the pre-review but not the final review
2. add the JOSS doi to the header
3. update the label to say joss accepted (in addition to pyos accepted!)

many thanks y'all!!
