### 2024-06-11 #146 @tomalrussell
Hey @lwasser thanks for checking in!

The package `v1.9.0` has been accepted into JOSS, with review issue https://github.com/openjournals/joss-reviews/issues/6490

DOI updated in the issue header: https://doi.org/10.21105/joss.06490
